% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geographical.R
\name{calculate_cell_area}
\alias{calculate_cell_area}
\title{Calculate area extent of grid cells}
\usage{
calculate_cell_area(
  x,
  grid,
  crs = sf::st_crs(x),
  tol = sqrt(.Machine[["double.eps"]])
)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{grid}{A \code{\link[raster:Raster-class]{raster::Raster}} or
a \code{stars::stars} object for coordinate system and cells.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-class]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{tol}{A numerical value to inform the site to gridcell matching.}
}
\value{
A \code{data.frame} with four columns: \var{\dQuote{Longitude}},
  \var{\dQuote{Latitude}}, \var{\dQuote{km2}}, and \var{\dQuote{rel}};
  and one row for each site of \code{sites}.
  Cell area on ellipsoid, based on \code{grid}, in square-kilometer
  \code{km2} and as fraction of maximal cell area on the equator \code{rel}.
}
\description{
Calculate area extent of grid cells
}
\examples{
r <- raster::raster(
  xmn = 0, xmx = 1,
  ymn = -90, ymx = 90,
  crs = "OGC:CRS84",
  resolution = c(1, 1)
)
n <- prod(dim(r))
r[] <- seq_len(n)
xy <- raster::sampleRegular(r, size = n, sp = TRUE)

## Calculate area for a subset of cells in grid
cell_areas <- calculate_cell_area(xy, grid = r)
cell_areas <- calculate_cell_area(xy, grid = stars::st_as_stars(r))

## Calculate are for all cells in grid
cell_areas2 <- calculate_cell_area(grid = r)


## Visualize cell area by latitude
with(cell_areas, graphics::plot(y, km2, type = "l"))

## Comparison with a spherical Earth
# A spherical Earth underestimates cell areas
# at mid latitudes compared to cell areas on a WGS84 ellipsoid as here
rel_spherical <- cos(cell_areas[, "y"] * pi / 180)
graphics::plot(
  abs(cell_areas[, "y"]),
  max(cell_areas[, "km2"]) * (cell_areas[, "rel"] - rel_spherical),
  pch = 46,
  xlab = "abs(Latitude)",
  ylab = "Cell area difference\n(ellipsoid - sphere; km2)"
)

}
\seealso{
\code{\link[geosphere]{areaPolygon}}
}
