% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geographical.R
\name{calculate_nominal_resolution}
\alias{calculate_nominal_resolution}
\title{Calculate "nominal resolution" of a grid}
\usage{
calculate_nominal_resolution(grid, maskvalue = NA)
}
\arguments{
\item{grid}{A \code{\link[terra:SpatRaster-class]{terra::rast}},
a \code{\link[raster:Raster-class]{raster::Raster}} or
a \code{stars::stars} object.
Gridcells with values that are not equal to a
\code{maskvalue} are included in the "nominal resolution" calculation.}

\item{maskvalue}{A vector. Values to mask out from \code{grid}.}
}
\value{
A character string with a "nominal resolution" value in kilometers.
}
\description{
Calculate "nominal resolution" of a grid
}
\examples{
r1 <- terra::rast(
  xmin = -120, xmax = -90,
  ymin = 30, ymax = 50,
  crs = "OGC:CRS84",
  resolution = c(0.5, 0.5)
)
xy <- terra::spatSample(r1, size = 200L, as.points = TRUE)
r1[xy] <- 1

calculate_nominal_resolution(r1)
calculate_nominal_resolution(stars::st_as_stars(r1))

r2 <- terra::rast(
  xmin = -2480000, xmax = 90000,
  ymin = 650000, ymax = 4020000,
  crs = "EPSG:6350",
  resolution = c(10000, 10000)
)
xy <- terra::spatSample(r2, size = 200L, as.points = TRUE)
r2[xy] <- 1

calculate_nominal_resolution(r2)
calculate_nominal_resolution(stars::st_as_stars(r2))

}
\references{
\var{CMIP6 Global Attributes, DRS, Filenames, Directory Structure,
  and CV’s}, 10 September 2018 (v6.2.7).
  \href{https://docs.google.com/document/d/1h0r8RZr_f3-8egBMMh7aqLwy3snpD6_MrDz1q8n5XUk/edit#bookmark=id.ibeh7ad2gpdi}{Appendix 2: Algorithms for Defining the "nominal_resolution" Attribute}
}
