% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{create_example_netCDFs}
\alias{create_example_netCDFs}
\title{Create example \var{netCDFs} for use by package}
\usage{
create_example_netCDFs(
  path,
  data_str = c("xyzt", "xyt", "xyz", "xy", "szt", "st", "sz", "s"),
  type_timeaxis = c("timeseries", "climatology"),
  overwrite = FALSE
)
}
\arguments{
\item{path}{A character string. The path to where the files will be
written to disk.}

\item{data_str}{A character string describing the dimensions of \code{data}
where \var{"xy"} stands for \var{x} and \var{y} spatial dimensions
if the spatial structure is gridded,
while \var{"s"} stands for \var{site} if the spatial structure are
discrete points;
\var{z} stands for a vertical dimension; and \var{t} stands for a
temporal dimension.}

\item{type_timeaxis}{A character string. Describing if the time dimension
represents a time series or a climatological time.}

\item{overwrite}{A logical value. If \code{TRUE}, file will be overwritten
if it already exists.}
}
\value{
A named list, invisibly, of full paths to the created files on disk.
}
\description{
This function creates different \var{netCDFs} for all requested conditions:
\itemize{
  \item \code{data_str}
  \item \code{type_timeaxis}
}
}
\examples{
tmp_nc <- create_example_netCDFs(tempdir(), "xyt", "timeseries")
data_xyt <- read_netCDF(
  tmp_nc[["xyt"]],
  method = "array",
  xy_names = c("x", "y")
)

if (requireNamespace("graphics")) {
  graphics::persp(
    x = data_xyt[["xyspace"]][["x"]],
    y = data_xyt[["xyspace"]][["y"]],
    z = data_xyt[["data"]][, , 15],
    theta = 30,
    phi = 30,
    expand = 0.5,
    col = "lightblue"
  )
}

unlink(unlist(tmp_nc))

}
