% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{get_data_dims}
\alias{get_data_dims}
\title{\var{XYZT} or \var{SZT} data dimensions for interacting with \var{netCDFs}}
\usage{
get_data_dims(
  data_str = c("xyzt", "xyt", "xyz", "xy", "szt", "st", "sz", "s"),
  dims = NULL
)
}
\arguments{
\item{data_str}{A character string describing the dimensions of \code{data}
where \var{"xy"} stands for \var{x} and \var{y} spatial dimensions
if the spatial structure is gridded,
while \var{"s"} stands for \var{site} if the spatial structure are
discrete points;
\var{z} stands for a vertical dimension; and \var{t} stands for a
temporal dimension.}

\item{dims}{An integer vector.}
}
\value{
A named integer vector with six elements for
  \describe{
    \item{ns}{sites, for a spatial representation of data as sites}
    \item{nx}{x coordinate, for a gridded spatial representation}
    \item{ny}{y coordinate, for a gridded spatial representation}
    \item{nz}{vertical levels}
    \item{nt}{temporal units}
  }
}
\description{
\var{XYZT} or \var{SZT} data dimensions for interacting with \var{netCDFs}
}
\examples{
x <- array(dim = c(17, 15, 12, 100))
get_data_dims("xyzt", dim(x))
get_data_dims("xyzt", c(17, 15, 12, 100))
get_data_dims("xyt", c(17, 15, 100))
get_data_dims("xyz", c(17, 15, 12))
get_data_dims("xy", c(17, 15))
get_data_dims("xy", c(17, 15, 3))

get_data_dims("szt", c(17 * 15, 12, 100))
get_data_dims("xy", c(17 * 15))
get_data_dims("xy", c(17 * 15, 3))

}
