% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_raster.R
\name{get_isoline_polygon}
\alias{get_isoline_polygon}
\title{Creates an isoline-polygon from data where values are larger than alpha}
\usage{
get_isoline_polygon(x, site_locations, grid, subset, alpha)
}
\arguments{
\item{x}{A numeric vector}

\item{site_locations}{An object that described geographic locations of
\code{data} that can be sent to \code{\link[rSW2st]{as_points}}}

\item{grid}{A \var{Raster*} object used as template.}

\item{subset}{A logical vector of length equal to \code{x}}

\item{alpha}{A numeric value}
}
\value{
A \code{\link[sp]{SpatialPolygons}} object.
}
\description{
Function converts data to a raster object and then draws a
polygon around those gridcells with a value larger than alpha
}
