% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{get_xyspace}
\alias{get_xyspace}
\title{Extract \var{xy-space} values}
\usage{
get_xyspace(
  x,
  crs,
  res,
  xy_names = c("lon", "lat"),
  tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{x}{An object that describes a gridded or discrete \var{xy-space}.
 Regular, rectangular grids are the only currently supported grid.
 This can be \itemize{
   \item an object identifying a \var{netCDF} file, i.e.,
         a character string as file name or
         an object of class \var{ncdf4} derived from \code{ncdf4::nc_open}
   \item a \code{\link[raster:RasterLayer-class]{raster::RasterLayer}}
         object,
   \item a \code{stars::stars} object,
   \item a list, such as the one produced by \code{\link{get_xyspace}}.
   \item an object with coordinate values for all \var{gridcell} centers
         that can be passed to \code{\link{as_points}};
}}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-class]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{res}{A numeric vector of length two. The (absolute values of)
cell size/resolution/delta of the \var{x} and \var{y} dimensions
in units of the \var{crs}.}

\item{xy_names}{A vector with two character strings. The names of the
\var{x} and \var{y} spatial dimensions of a \var{netCDF} file.}

\item{tol}{A numeric value. The tolerance applied to determine if a grid
is regular and to calculate grid resolution.}
}
\value{
A list with three elements:
 two vectors one each containing all unique values of the
 \var{x} coordinate and the \var{y} coordinates of
 (i) all \var{gricell} centers, if gridded, or
 (ii) all \var{sites}, if discrete;
 and a vector \var{res} with the (regular) \var{x} and \var{y} resolutions,
 if gridded (NA, if discrete).
}
\description{
Extract \var{xy-space} values
}
\section{Notes}{

The argument \code{crs} is only used if \code{grid} is a \code{data.frame},
in which case it is passed to \code{\link{as_points}}.
Otherwise, it is ignored and can be missing.


The argument \code{res} is only used in two cases: \itemize{
  \item if \code{grid} is a list,
        such as the one produced by \code{\link{get_xyspace}}, but does
        not contain a named element \var{res}; or
  \item if \code{grid} is an object with coordinate values for
        all \var{gridcell} centers.
}
Otherwise, it is ignored and can be missing.
}

\examples{
# grid as raster object
r <- raster::raster(
  xmn = 0, xmx = 120,
  ymn = 0, ymx = 45,
  crs = "OGC:CRS84",
  resolution = c(1, 1)
)
get_xyspace(r)

# grid as data frame with coordinate values
rdf <- raster::coordinates(r)
get_xyspace(rdf, crs = "OGC:CRS84", res = c(1, 1))

# a list with vectors for all x values and all y values (and resolution)
rl <- list(
  x = sort(unique(rdf[, 1])),
  y = sort(unique(rdf[, 2]))
)
get_xyspace(c(rl, list(res = c(1, 1))))
get_xyspace(rl, res = c(1, 1))

}
