% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_raster.R
\name{isoline_from_raster}
\alias{isoline_from_raster}
\title{Polygon around \var{gridcells} with a value larger than or equal to the
specified threshold}
\usage{
isoline_from_raster(grid, alpha)
}
\arguments{
\item{grid}{A two-dimensional
\code{\link[raster:RasterLayer-class]{raster::RasterLayer}},
\code{\link[terra:SpatRaster-class]{terra::SpatRaster}}, or
\code{stars} object.}

\item{alpha}{A numeric value. Threshold value.}
}
\value{
A \code{sf} polygon
}
\description{
Polygon around \var{gridcells} with a value larger than or equal to the
specified threshold
}
\examples{
r <- stars::st_as_stars(
  data.frame(
    x = rep(1:10, times = 10),
    y = rep(10:1, each = 10),
    value = 1:100
  ),
  dims = c("x", "y")
)

ip1 <- isoline_from_raster(r, alpha = 45)
ip2 <- isoline_from_raster(r, alpha = 87)

if (requireNamespace("grDevices")) {
  plot(r, reset = FALSE)
  plot(ip1, col = NA, border = "black", add = TRUE)
  plot(ip2, col = NA, border = "blue", add = TRUE)
  grDevices::dev.off()
}

}
