% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{ncDataType}
\alias{ncDataType}
\title{Translate to \code{netCDF} data types}
\usage{
ncDataType(dataType, stopOnError = TRUE)
}
\arguments{
\item{dataType}{A character string. A \code{netCDF} data type, see \code{\link[=ncDataType]{ncDataType()}}.}

\item{stopOnError}{A logical value.}
}
\value{
A standardized \code{netCDF} library data type.
If not found, then return value depends on \code{stopOnError}:
\itemize{
\item if \code{TRUE}, then signal an error
\item if \code{FALSE}, then return unmodified value of \code{dataType}
}
}
\description{
Translate to \code{netCDF} data types
}
\examples{
ncDataType("integer")
ncDataType("noType", stopOnError = FALSE)

}
