% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{populate_netCDF}
\alias{populate_netCDF}
\title{Add data to an existing \var{netCDF} file}
\usage{
populate_netCDF(
  filename,
  data,
  var_names,
  data_str = c("xyzt", "xyt", "xyz", "xy", "szt", "st", "sz", "s"),
  nc_name_crs = "crs",
  nc_name_crs_wkt = "crs_wkt",
  xy_names = c("lon", "lat"),
  time_ids = -1,
  vertical_ids = -1,
  ...
)
}
\arguments{
\item{filename}{A character string. The name of the \var{netCDF} file.}

\item{data}{A numeric array or vector (optional). A vector is converted
to a one-column matrix.}

\item{var_names}{A character vector of strings. The \var{netCDF} variable
names.}

\item{data_str}{A character string describing the dimensions of \code{data}
where \var{"xy"} stands for \var{x} and \var{y} spatial dimensions
if the spatial structure is gridded,
while \var{"s"} stands for \var{site} if the spatial structure are
discrete points;
\var{z} stands for a vertical dimension; and \var{t} stands for a
temporal dimension.}

\item{nc_name_crs}{A character string. The name of the \var{crs} variable
in the \var{netCDF}.
Function \code{\link{create_netCDF}} hard codes \var{"crs"}.}

\item{nc_name_crs_wkt}{A character string. The name of the attribute
that holds the \var{WKT2} string of the \var{crs} variable
in the \var{netCDF}.
Function \code{\link{create_netCDF}} hard codes \var{"crs_wkt"}.}

\item{xy_names}{A vector with two character strings. The names of the
\var{x} and \var{y} spatial dimensions of a \var{netCDF} file.}

\item{time_ids}{An integer vector. The index to read a subset of
time steps; a value of \code{-1} means to read all.}

\item{vertical_ids}{An integer vector. The index to read a subset of
vertical steps; a value of \code{-1} means to read all.}

\item{...}{Additional arguments passed on to the specific functions.}
}
\description{
Add data to an existing \var{netCDF} file
}
\section{Notes}{
 This function is not yet implemented.
}

