% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{read_crs_from_netCDF}
\alias{read_crs_from_netCDF}
\title{Read \var{crs} projection from a \var{netCDF}}
\usage{
read_crs_from_netCDF(x, nc_name_crs = "crs", nc_name_crs_wkt = "crs_wkt")
}
\arguments{
\item{x}{An object identifying a \var{netCDF} file, i.e.,
a character string as file name, an object of class \var{"NetCDF"},
or an object of class \var{"ncdf4"} (which may be closed, with a warning,
and the file re-opened by its file name).}

\item{nc_name_crs}{A character string. The name of the \var{crs} variable
in the \var{netCDF}.
Function \code{\link{create_netCDF}} hard codes \var{"crs"}.}

\item{nc_name_crs_wkt}{A character string. The name of the attribute
that holds the \var{WKT2} string of the \var{crs} variable
in the \var{netCDF}.
Function \code{\link{create_netCDF}} hard codes \var{"crs_wkt"}.}
}
\value{
An object of the \var{crs}-class of package \pkg{sf}.
}
\description{
Read \var{crs} projection from a \var{netCDF}
}
\section{Details}{

Example code is available in the documentation of
\code{\link{read_netCDF}}.
}

