% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{read_netCDF_to_array}
\alias{read_netCDF_to_array}
\title{Read data from a netCDF into an array or matrix}
\usage{
read_netCDF_to_array(x, locations)
}
\arguments{
\item{x}{a netCDF file}
}
\value{
an array
}
\description{
Read data from a netCDF into an array or matrix
}
