% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSW2st-package.R
\docType{package}
\name{rSW2st}
\alias{rSW2st-package}
\alias{rSW2st}
\title{Package \pkg{rSW2st}: Collection of general purpose functions that operate on
spatially and temporally explicit objects
to support \pkg{SOILWAT2} and \pkg{STEPWAT2} simulation experiments.}
\description{
Collection of general purpose functions that operate on spatially and temporally explicit objects to support SOILWAT2/STEPWAT2 simulation experiments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DrylandEcology/rSW2st}
  \item Report bugs at \url{https://github.com/DrylandEcology/rSW2st/issues}
}

}
\author{
\strong{Maintainer}: Daniel Schlaepfer \email{danielrschlaepfer@gmail.com} (\href{https://orcid.org/0000-0001-9973-2065}{ORCID})

Authors:
\itemize{
  \item Daniel Schlaepfer \email{danielrschlaepfer@gmail.com} (\href{https://orcid.org/0000-0001-9973-2065}{ORCID})
}

}
