% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{rSW2st_crs}
\alias{rSW2st_crs}
\alias{crs_units}
\title{Determines the coordinate reference system}
\usage{
crs_units(crs)
}
\arguments{
\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-class]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}
}
\value{
The function \code{crs_units} returns the \var{crs} units
  as a character string or \code{NA}.
}
\description{
Determines the coordinate reference system
}
\examples{
crs_units(4326)
crs_units("EPSG:4326")
crs_units(sp::CRS("+init=EPSG:4326"))

}
\seealso{
\code{\link[sf]{st_crs}}
}
