% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{rSW2st_netCDF}
\alias{rSW2st_netCDF}
\alias{netCDF}
\alias{netcdf}
\title{Interaction of \var{rSW2} objects with \var{netCDFs}}
\arguments{
\item{data}{A numeric array or vector (optional). A vector is converted
to a one-column matrix.}

\item{grid}{An object that describes a gridded \var{xy-space}.
Regular, rectangular grids are the only currently supported type.
This can be \itemize{
  \item a \code{\link[raster:RasterLayer-class]{raster::RasterLayer}}
        object,
  \item a \code{stars::stars} object,
  \item a file name pointing to such a raster on disk;
  \item a list, such as the one produced by \code{\link{xy_from_grid}}.
  \item an object with coordinate values for all \var{gridcell} centers
        that can be passed to \code{\link{as_points}};
}
The \var{crs} of the grid coordinate values must match the one of the
data locations.}

\item{locations}{An object from which \var{x} and \var{y} coordinate values
can be extracted that describe the \var{xy} locations of each \code{data}
row, e.g., a matrix or \var{data.frame} or a spatial points object
inheriting from \var{sf} or \var{Spatial*}.}

\item{locations_crs}{An object which is a \var{crs} or
from which one can be derived
that describes the \var{crs} of \code{locations}.}

\item{data_str}{A character string describing the dimensions of \code{data}
where \var{"xy"} stands for \var{x} and \var{y} spatial dimensions
if the spatial structure is gridded,
while \var{"s"} stands for \var{site} if the spatial structure are
discrete points;
\var{z} stands for a vertical dimension; and \var{t} stands for a
temporal dimension.}
}
\description{
Interaction of \var{rSW2} objects with \var{netCDFs}
}
