% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setAxisBoundsNCSW}
\alias{setAxisBoundsNCSW}
\alias{setSpatialBoundsNCSW}
\title{Set bounds for a coordinate variable}
\usage{
setAxisBoundsNCSW(
  x,
  nameBndsVar,
  nameDim,
  valuesBnds = NULL,
  calculateValuesBndsIfMissing = FALSE,
  boundsAttributeName = "bounds",
  nameBndsDim = "bnds"
)

setSpatialBoundsNCSW(
  x,
  nameBndsDim = "bnds",
  nameBndsX = "lon_bnds",
  nameBndsY = "lat_bnds",
  nameDimX = "longitude",
  nameDimY = "latitude",
  valuesBndsX = NULL,
  valuesBndsY = NULL
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed),
an object of class \code{"NetCDF"} from the \code{RNetCDF} package (kept open), or
an object of class \code{"ncdf4"} from the \code{ncdf4} package (closed with a warning;
the file is then opened by its file name and closed on exit).}

\item{nameBndsVar}{A character string. The name of the selected axis bounds
variable.}

\item{nameDim}{A character string. The name of the associated dimensions.}

\item{valuesBnds}{A numeric vector. The values of the selected axis bounds
variable.}

\item{calculateValuesBndsIfMissing}{A logical value. The values of
the selected axis bounds will derived from the associated axis variable
assuming symmetric bounds.}

\item{boundsAttributeName}{A character vector. The name of the bounds
attribute that is added to the \code{nameDim} variable. In most cases, this
attribute's name is \code{"bounds"} or \code{"climatology"}.}

\item{nameBndsDim}{A character string. The name of the bounds dimension.}

\item{nameBndsX}{A character string. The name of the \code{X}-axis bounds
variable.}

\item{nameBndsY}{A character string. The name of the \code{Y}-axis bounds
variable.}

\item{nameDimX}{A character string. The name of the \code{X}-axis dimension and
coordinate variable.}

\item{nameDimY}{A character string. The name of the \code{Y}-axis dimension and
coordinate variable.}

\item{valuesBndsX}{A numeric vector. The values of the \code{X}-axis bounds
variable. Values will derived from \code{X}-axis variable assuming
symmetric bounds if \code{NULL}.}

\item{valuesBndsY}{A numeric vector. The values of the \code{Y}-axis bounds
symmetric bounds if \code{NULL}.}
}
\value{
\code{\link[=setAxisBoundsNCSW]{setAxisBoundsNCSW()}}: set bounds on a selected axis.

\code{\link[=setSpatialBoundsNCSW]{setSpatialBoundsNCSW()}}: set spatial bounds on x-axis and y-axis.
}
\description{
Set bounds for a coordinate variable
}
