% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setAxisMonthClimatologyNCSW}
\alias{setAxisMonthClimatologyNCSW}
\title{Add a monthly climatology dimension and coordinate variable}
\usage{
setAxisMonthClimatologyNCSW(
  x,
  startYear,
  endYear,
  nameAxis = "time",
  dataType = "NC_DOUBLE"
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{dataType}{A character string. A \code{netCDF} data type.}
}
\description{
Add a monthly climatology dimension and coordinate variable
}
