% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setAxisNCSW}
\alias{setAxisNCSW}
\alias{setAxisSiteNCSW}
\alias{setAxisVerticalNCSW}
\alias{setAxisPFTsNCSW}
\alias{setAxisTimeNCSW}
\alias{setAxisMonthClimatologyNCSW}
\title{Add a dimension and associated (coordinate) variable}
\usage{
setAxisNCSW(
  x,
  nameAxis,
  dataType,
  values = NULL,
  isUnlimitedDim = FALSE,
  axis = NULL,
  long_name = NULL,
  units = NULL,
  attributes = NULL
)

setAxisSiteNCSW(
  x,
  siteValues,
  nameAxis = "site",
  units = "1",
  dataType = "NC_INT"
)

setAxisVerticalNCSW(
  x,
  verticalValues,
  verticalUpperBound = NULL,
  verticalLowerBound = NULL,
  verticalType = c("values", "layers"),
  nameAxis = "vertical",
  units = "cm",
  dataType = "NC_INT"
)

setAxisPFTsNCSW(
  x,
  pftValues = c(Tree = "Trees", Shrub = "Shrubs", Forb = "Forbs", Grass = "Grasses"),
  nameAxis = "pft",
  dataType = c("NC_STRING", "NC_BYTE")
)

setAxisTimeNCSW(
  x,
  timeValues,
  startYear = NULL,
  timeUnits = NULL,
  nameAxis = "time",
  calendar = c("standard", "365_day", "366_day"),
  dataType = "NC_DOUBLE",
  isUnlimitedDim = FALSE
)

setAxisMonthClimatologyNCSW(
  x,
  startYear,
  endYear,
  nameAxis = "time",
  dataType = "NC_DOUBLE"
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed),
an object of class \code{"NetCDF"} from the \code{RNetCDF} package (kept open), or
an object of class \code{"ncdf4"} from the \code{ncdf4} package (closed with a warning;
the file is then opened by its file name and closed on exit).}

\item{nameAxis}{A character string. The name of the axis dimension and
associated coordinate variable.}

\item{dataType}{A character string. A \code{netCDF} data type, see \code{\link[=ncDataType]{ncDataType()}}.}

\item{values}{A numeric object. The values to write a variable if not \code{NULL}.}

\item{isUnlimitedDim}{A logical value. Define the new dimension as unlimited?}

\item{axis}{A character string. The \code{"axis"} attribute.}

\item{long_name}{A character string. The \code{"long_name"} attribute.}

\item{units}{A character string. The length units of \code{verticalType}.}

\item{attributes}{A named vector or named list of character strings.}

\item{siteValues}{A numeric object. The values of the discrete \code{site} axis
variable.}

\item{verticalValues}{A numeric vector. The values of the vertical axis
variable.}

\item{verticalUpperBound}{A numerical vector. The upper (top) soil depths
(bounds) of each soil layer (as defined by \code{verticalValues}).}

\item{verticalLowerBound}{A numerical vector. The lower (bottom) soil depths
(bounds) of each soil layer (as defined by \code{verticalValues}).}

\item{verticalType}{A character string. The type of the vertical axis.}

\item{pftValues}{A vector of character strings. The value of the variable
representing plant functional types.}

\item{timeValues}{A numeric vector. The values of the time axis variable.}

\item{startYear}{An integer value. The first calendar year of the period
which the climatology represents.}

\item{timeUnits}{A character string. The time units,
e.g., \verb{"days since ...}. \code{timeUnits} takes precedence over \code{startYear}
if both are provided.}

\item{calendar}{A character string representing the calendar type.}

\item{endYear}{An integer value. The last calendar year of the period
which the climatology represents.}
}
\value{
\code{\link[=setAxisSiteNCSW]{setAxisSiteNCSW()}} adds a station/site dimension and associated
coordinate variable.

\code{\link[=setAxisVerticalNCSW]{setAxisVerticalNCSW()}} adds a vertical dimension and associated
coordinate variable (\code{Z}-axis).

\code{\link[=setAxisPFTsNCSW]{setAxisPFTsNCSW()}} adds a plant functional type dimension and
associated coordinate variable.

\code{\link[=setAxisTimeNCSW]{setAxisTimeNCSW()}} adds a time dimension and associated
coordinate variable (\code{T}-axis).

\code{\link[=setAxisMonthClimatologyNCSW]{setAxisMonthClimatologyNCSW()}} adds a monthly climatology dimension
and associated coordinate variable.
}
\description{
Add a dimension and associated (coordinate) variable
}
