% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setAxisPFTsNCSW}
\alias{setAxisPFTsNCSW}
\title{Add a plant functional type dimension and coordinate variable}
\usage{
setAxisPFTsNCSW(
  x,
  pftValues = c(Tree = "Trees", Shrub = "Shrubs", Forb = "Forbs", Grass = "Grasses"),
  nameAxis = "pft",
  dataType = c("NC_STRING", "NC_BYTE")
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{pftValues}{A vector of character strings. The value of the variable
representing plant functional types.}

\item{nameAxis}{A character string. The name of the axis dimension and
associated coordinate variable.}

\item{dataType}{A character string. A \code{netCDF} data type, see \code{\link[=ncDataType]{ncDataType()}}.}
}
\description{
Add a plant functional type dimension and coordinate variable
}
