% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setAxisSiteNCSW}
\alias{setAxisSiteNCSW}
\title{Add a station/site dimension and coordinate variable}
\usage{
setAxisSiteNCSW(
  x,
  siteValues,
  nameAxis = "site",
  units = "1",
  dataType = "NC_INT"
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{nameAxis}{A character string. The name of the axis dimension and
associated coordinate variable.}

\item{dataType}{A character string. A \code{netCDF} data type, see \code{\link[=ncDataType]{ncDataType()}}.}
}
\description{
Add a station/site dimension and coordinate variable
}
