% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setCRSNCSW}
\alias{setCRSNCSW}
\alias{setCRSWGS84NCSW}
\title{Create a \code{CRS} variable}
\usage{
setCRSNCSW(
  x,
  nameCRS = "crs",
  grid_mapping_name = NULL,
  crs_wkt = NULL,
  attributes = NULL
)

setCRSWGS84NCSW(x, nameCRS = "crs")
}
\arguments{
\item{x}{A character string (file name that will be opened and closed),
an object of class \code{"NetCDF"} from the \code{RNetCDF} package (kept open), or
an object of class \code{"ncdf4"} from the \code{ncdf4} package (closed with a warning;
the file is then opened by its file name and closed on exit).}

\item{nameCRS}{A character string. The name of the \code{CRS} variable.}

\item{grid_mapping_name}{A character string. The \code{"grid_mapping_name"}
attribute.}

\item{crs_wkt}{A character string. The \code{"crs_wkt"} attribute.}

\item{attributes}{A named vector or named list of character strings.}
}
\value{
\code{\link[=setCRSNCSW]{setCRSNCSW()}}: creates a \code{CRS} variable with custom attributes

\code{\link[=setCRSWGS84NCSW]{setCRSWGS84NCSW()}}: creates a \code{CRS} variable with \code{WGS84} attributes
}
\description{
Create a \code{CRS} variable
}
\examples{
fn <- tempfile(fileext = ".nc")
nc <- RNetCDF::create.nc(fn, format = "netcdf4")
setCRSWGS84NCSW(nc)

RNetCDF::print.nc(nc)
RNetCDF::close.nc(nc)
unlink(fn)

}
