% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setCRSWGS84NCSW}
\alias{setCRSWGS84NCSW}
\title{Create a \code{CRS} variable with \code{WGS84} attributes}
\usage{
setCRSWGS84NCSW(x, nameCRS)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{nameCRS}{A character string. The name of the \code{CRS} variable.}
}
\description{
Create a \code{CRS} variable with \code{WGS84} attributes
}
\examples{
fn <- tempfile(fileext = ".nc")
nc <- RNetCDF::create.nc(fn, format = "netcdf4")
setCRSWGS84NCSW(nc, nameCRS = "crs")

RNetCDF::print.nc(nc)
RNetCDF::close.nc(nc)
unlink(fn)

}
