% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setGlobalAttributesNCSW}
\alias{setGlobalAttributesNCSW}
\title{Set global attributes of type character}
\usage{
setGlobalAttributesNCSW(x, attributes)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed),
an object of class \code{"NetCDF"} from the \code{RNetCDF} package (kept open), or
an object of class \code{"ncdf4"} from the \code{ncdf4} package (closed with a warning;
the file is then opened by its file name and closed on exit).}

\item{attributes}{A named vector or named list of character strings.}
}
\description{
Set global attributes of type character
}
\examples{
fn <- tempfile(fileext = ".nc")
nc <- RNetCDF::create.nc(fn)
setGlobalAttributesNCSW(nc, c(featureType = "timeSeries", frequency = "day"))
setGlobalAttributesNCSW(nc, c(source = "SOILWAT2v8.2.0"))

RNetCDF::print.nc(nc)
RNetCDF::close.nc(nc)
unlink(fn)

}
