% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setVariableNCSW}
\alias{setVariableNCSW}
\title{Add a variable}
\usage{
setVariableNCSW(
  x,
  varName,
  values = NULL,
  start = NA,
  count = NA,
  var_chunksizes_xyzt = NA,
  dataType = "NC_DOUBLE",
  dimensions = NULL,
  deflate = 5L,
  shuffle = !anyNA(deflate),
  long_name = NULL,
  units = NULL,
  cell_method = NULL,
  cell_methods = cell_method,
  coordinates = paste("latitude", "longitude"),
  grid_mapping = "crs",
  attributes = NULL,
  addFillValue = TRUE
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed),
an object of class \code{"NetCDF"} from the \code{RNetCDF} package (kept open), or
an object of class \code{"ncdf4"} from the \code{ncdf4} package (kept open).}

\item{varName}{A character string. The name of the variable.
see \code{\link[RNetCDF:var.put.nc]{RNetCDF::var.put.nc()}} for more detail.}

\item{values}{A numeric object. The values to write a variable if not \code{NULL}.}

\item{start}{A numeric vector. The start indices at which to write values,
see \code{\link[RNetCDF:var.put.nc]{RNetCDF::var.put.nc()}} for more detail.}

\item{count}{A numeric vector. The counts of writing values,
see \code{\link[RNetCDF:var.put.nc]{RNetCDF::var.put.nc()}} for more detail.}

\item{var_chunksizes_xyzt}{A numeric vector. The chunk sizes if not \code{NA},
see \code{\link[RNetCDF:var.def.nc]{RNetCDF::var.def.nc()}} for more detail.}

\item{dataType}{A character string. A \code{netCDF} data type, see \code{\link[=ncDataType]{ncDataType()}}.}

\item{dimensions}{A vector, see \code{\link[RNetCDF:var.def.nc]{RNetCDF::var.def.nc()}} for more detail.}

\item{deflate}{A numeric value or \code{NA},
see \code{\link[RNetCDF:var.def.nc]{RNetCDF::var.def.nc()}} for more detail.}

\item{shuffle}{A logical value. Turn on the shuffle filter?
Default is to turn it on if \code{deflate} is not \code{NA};
see \code{\link[RNetCDF:var.def.nc]{RNetCDF::var.def.nc()}} for more detail.}

\item{long_name}{A character string. The \code{"long_name"} attribute.}

\item{units}{A character string.  The \code{"units"} attribute.}

\item{cell_method}{A character string. Passed to \code{"cell_methods"} for
backwards compatibility; will be removed in the version.}

\item{cell_methods}{A character string. The \code{"cell_methods"} attribute.
Previously, misspelled as \code{"cell_method"}.}

\item{coordinates}{A character string. The \code{"coordinates"} attribute.}

\item{grid_mapping}{A character string. The \code{"grid_mapping"} attribute.}

\item{attributes}{A named vector or named list of character strings.}

\item{addFillValue}{A logical value. Add a \code{"_FillValue"} attribute? The
value is determined by \code{dataType} and \code{\link[=fillValue]{fillValue()}}.}
}
\description{
Add a variable
}
\section{Details}{

\enumerate{
\item Create variable of \code{dataType} and \code{dimensions} if not present,
see \code{\link[RNetCDF:var.def.nc]{RNetCDF::var.def.nc()}}.
\item Write values if \code{values} are not \code{NULL}
(using \code{count} and \code{start} if provided), see \code{\link[RNetCDF:var.put.nc]{RNetCDF::var.put.nc()}}.
\item Add attributes \code{"long_name"}, \code{"units"}, \code{"cell_methods"},
\code{"coordinates"} and \code{"grid_mapping"} if provided.
If \code{"long_name"} is not provided and there is no attribute with that
name, then \code{varName} is used as value for \code{"long_name"}.
\item Add additional attributes if provided.
}
}

