% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{uniqueDifferences}
\alias{uniqueDifferences}
\title{Unique differences between adjacent values}
\usage{
uniqueDifferences(
  x,
  uniqueDifferences = TRUE,
  allowMultiple = FALSE,
  tolerance = sqrt(.Machine[["double.eps"]])
)
}
\arguments{
\item{x}{A numeric vector.}

\item{uniqueDifferences}{A logical value. Find unique differences between
adjacent values in \code{x}.}

\item{allowMultiple}{A logical value. Allow more than one unique difference.}

\item{tolerance}{A numeric value. Tolerance to identify unique differences.}
}
\description{
Check if values are regularly spaced.
}
\examples{
x <- seq(0, 1, by = 1 / 24)
uniqueDifferences(x)

x[[2L]] <- x[[2L]] + sqrt(.Machine[["double.eps"]])
try(uniqueDifferences(x))
uniqueDifferences(x, tolerance = 1e-7)
uniqueDifferences(x, allowMultiple = TRUE)

}
