% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crs.R
\name{utm_zone}
\alias{utm_zone}
\title{\var{UTM} zone based on geographic location(s)}
\usage{
utm_zone(x, crs = 4326)
}
\arguments{
\item{x}{A numerical two-dimensional object
(a \code{matrix}, \code{array}, or \code{data.frame})
with longitude/X, latitude/Y as columns;
a \code{\link[sp:SpatialPoints-class]{sp::SpatialPoints}} object; or
a \var{sf} object with a point geometry,
i.e., an object with a class \var{sf} or \var{sfc}.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-class]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}
}
\value{
A list with two elements: \describe{
  \item{utm_zone}{The UTM zone number as integer value.}
  \item{utm_NS}{North/South indicator as character "N" or "S"}.
}
}
\description{
The function determines the \var{UTM} number and south/north location
for the mid-point of \code{x}.
}
\examples{
locations <- matrix(
  data = c(-120.325, -111.245, 39.855, 36.753),
  nrow = 2
)

utm_zone(locations)
utm_zone(locations[1, ])
utm_zone(locations[2, ])

}
\references{
Convert Latitude/Longitude to UTM
  \url{https://www.wavemetrics.com/code-snippet/convert-latitudelongitude-utm}
  (attributed to Chuck Gantz).
}
