% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{xy_from_grid}
\alias{xy_from_grid}
\title{Extract unique \var{x} and \var{y} spanning a gridded \var{xy-space}}
\usage{
xy_from_grid(grid, crs, res)
}
\arguments{
\item{grid}{An object that describes a gridded \var{xy-space}.
Regular, rectangular grids are the only currently supported type.
This can be \itemize{
  \item a \code{\link[raster:RasterLayer-class]{raster::RasterLayer}}
        object,
  \item a \code{stars::stars} object,
  \item a file name pointing to such a raster on disk;
  \item a list, such as the one produced by \code{\link{xy_from_grid}}.
  \item an object with coordinate values for all \var{gridcell} centers
        that can be passed to \code{\link{as_points}};
}
The \var{crs} of the grid coordinate values must match the one of the
data locations.}

\item{crs}{An object which is a \var{crs} or from which one can be derived.
\code{x} can
be numeric as a \var{EPSG} number;
a character string as a \var{wkt};
a character string as a \var{proj4} (not recommended because outdated);
or of a class including
  \code{\link[raster:Raster-class]{raster::Raster}},
  \code{\link[sp:Spatial-class]{sp::Spatial}},
  \code{\link[sp:CRS-class]{sp::CRS}},
  or a \code{\link[sf]{sf}} or \code{\link[sf]{sfc}} class.}

\item{res}{A numeric vector of length two. The (absolute values of)
cell size/resolution/delta of the \var{x} and \var{y} dimensions
in units of the \var{crs}.}
}
\value{
A list with three elements:
 two vectors one each containing all unique values of the
 \var{x} coordinate and the \var{y} coordinates of all \var{gricell} centers;
 and a vector \var{res} with the (regular) \var{x} and \var{y} resolutions.
}
\description{
Extract unique \var{x} and \var{y} spanning a gridded \var{xy-space}
}
\section{Notes}{

The argument \code{crs} is only used if \code{grid} is a \code{data.frame},
in which case it is passed to \code{\link{as_points}}.
Otherwise, it is ignored and can be missing.


The argument \code{res} is only used in two cases: \itemize{
  \item if \code{grid} is a list,
        such as the one produced by \code{\link{xy_from_grid}}, but does
        not contain a named element \var{res}; or
  \item if \code{grid} is an object with coordinate values for
        all \var{gridcell} centers.
}
Otherwise, it is ignored and can be missing.
}

\examples{
# grid as raster object
r <- raster::raster(
  xmn = 0, xmx = 120,
  ymn = 0, ymx = 45,
  crs = "OGC:CRS84",
  resolution = c(1, 1)
)
xy_from_grid(r)

# grid as data frame with coordinate values
rdf <- raster::coordinates(r)
xy_from_grid(rdf, crs = "OGC:CRS84", res = c(1, 1))

# a list with vectors for all x values and all y values (and resolution)
rl <- list(
  x = sort(unique(rdf[, 1])),
  y = sort(unique(rdf[, 2]))
)
xy_from_grid(c(rl, list(res = c(1, 1))))
xy_from_grid(rl, res = c(1, 1))

}
