% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_geographical.R
\name{calculate_nominal_resolution}
\alias{calculate_nominal_resolution}
\title{Calculate "nominal resolution" of grid}
\usage{
calculate_nominal_resolution(grid, sites, cell_areas_km2)
}
\arguments{
\item{grid}{A raster object}

\item{sites}{A raster object. A numeric array, matrix, or data.frame where
each row is a site and the columns contain values for long and lat.}

\item{cell_areas_km2}{A string of numeric values. Equal in length to the
length of sites. Area each cell represents in km2.}
}
\description{
Calculate "nominal resolution" of grid
}
\examples{
r <- raster::raster(
  xmn = -120, xmx = -90,
  ymn = 30, ymx = 50,
  crs ="+init=epsg:4326",
  resolution = c(0.5, 0.5)
)
r[] <- seq_len(prod(dim(r)))
xy <- raster::sampleRandom(r, size = 50, sp = TRUE)
grid_cell_area <- calculate_cell_area(sites = xy, grid = r)

calculate_nominal_resolution(
  grid = r,
  sites = xy,
  cell_areas_km2 = grid_cell_area[, "km2"]
)

}
\references{
CMIP6 Global Attributes, DRS, Filenames, Directory Structure,
  and CV’s 10 September 2018 (v6.2.7)
  Appendix 2: Algorithms for Defining the "nominal_resolution" Attribute
  \url{https://docs.google.com/document/d/1h0r8RZr_f3-8egBMMh7aqLwy3snpD6_MrDz1q8n5XUk/edit#bookmark=id.ibeh7ad2gpdi}
}
