% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{checkSpatialNCSW}
\alias{checkSpatialNCSW}
\title{Check spatial structure of a \code{netCDF}}
\usage{
checkSpatialNCSW(
  x,
  nameDimX = "longitude",
  nameDimY = "latitude",
  nameCRS = "crs",
  expectedSpatialDims = c(NA, NA),
  expectedSpatialExtent = c(NA, NA, NA, NA),
  tolerance = sqrt(.Machine[["double.eps"]])
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}
}
\description{
Check spatial structure of a \code{netCDF}
}
