% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{checkSpatialNCSW}
\alias{checkSpatialNCSW}
\title{Check spatial structure of a \code{netCDF}}
\usage{
checkSpatialNCSW(
  x,
  nameDimX = "longitude",
  nameDimY = "latitude",
  expectedSpatialDims = c(NA, NA),
  expectedSpatialExtent = c(xmin = NA, xmax = NA, ymin = NA, ymax = NA),
  tolerance = sqrt(.Machine[["double.eps"]])
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed),
an object of class \code{"NetCDF"} from the \code{RNetCDF} package (kept open), or
an object of class \code{"ncdf4"} from the \code{ncdf4} package (kept open).}

\item{nameDimX}{A character string. The name of the \code{X}-axis dimension and
coordinate variable.}

\item{nameDimY}{A character string. The name of the \code{Y}-axis dimension and
coordinate variable.}

\item{expectedSpatialDims}{A vector of length 2. Function will check
spatial dimensions of \code{x} against \code{expectedSpatialDims} if not \code{NA}.}

\item{expectedSpatialExtent}{A named vector of length 4. Function will check
spatial extent of \code{x} against \code{expectedSpatialExtent} if not \code{NA}.}

\item{tolerance}{A numeric value. Tolerance to identify unique differences.}
}
\description{
Check spatial structure of a \code{netCDF}
}
