% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{deleteGlobalAttributesNCSW}
\alias{deleteGlobalAttributesNCSW}
\title{Remove global attributes}
\usage{
deleteGlobalAttributesNCSW(x, attributes)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{attributes}{A named vector or named list of character strings.}
}
\description{
Remove global attributes
}
\examples{
fn <- tempfile(fileext = ".nc")
nc <- RNetCDF::create.nc(fn)
setGlobalAttributesNCSW(nc, c(featureType = "timeSeries", frequency = "day"))
deleteGlobalAttributesNCSW(nc, "featureType")

RNetCDF::print.nc(nc)
RNetCDF::close.nc(nc)
unlink(fn)

}
