% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{fillValue}
\alias{fillValue}
\title{Identify the fill value corresponding to a data type}
\usage{
fillValue(dataType)
}
\arguments{
\item{dataType}{A character string. A \code{netCDF} data type.}
}
\description{
Identify the fill value corresponding to a data type
}
\examples{
fillValue("integer")

}
