% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{get_nc_type}
\alias{get_nc_type}
\title{Guess \code{netCDF} data type of an object}
\usage{
get_nc_type(x)
}
\arguments{
\item{x}{An object}
}
\value{
A text string of supported \code{netCDF} data types. Supported
types are:
\code{"NC_INT"}, \code{"NC_DOUBLE"}, \code{"NC_STRING"}
}
\description{
Guess \code{netCDF} data type of an object
}
\examples{
get_nc_type("test") ## "NC_STRING"
get_nc_type(c(1L, 5L)) ## "NC_INT"
get_nc_type(1) ## "NC_DOUBLE"
\dontrun{get_nc_type(TRUE)} ## error

}
\seealso{
\code{\link[RNetCDF:var.def.nc]{RNetCDF::var.def.nc()}}
}
