% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{is_netCDF_gridded}
\alias{is_netCDF_gridded}
\title{Determine if a \var{netCDF} has a gridded or discrete \var{xy-space}}
\usage{
is_netCDF_gridded(x, xy_names = c("lon", "lat"))
}
\arguments{
\item{x}{An object identifying a \var{netCDF} file, i.e.,
a character string as file name or an object of class \var{ncdf4} derived
from \code{ncdf4::nc_open}.}

\item{xy_names}{A vector with two character strings. The names of the
\var{x} and \var{y} spatial dimensions of a \var{netCDF} file.}
}
\value{
A logical value
}
\description{
Determine if a \var{netCDF} has a gridded or discrete \var{xy-space}
}
