% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_raster.R
\name{isoline_from_raster}
\alias{isoline_from_raster}
\title{Polygon around \var{gridcells} with a value larger than or equal to the
specified threshold}
\usage{
isoline_from_raster(grid, alpha)
}
\arguments{
\item{grid}{A \code{\link[raster:RasterLayer-class]{raster::RasterLayer}}.}

\item{alpha}{A numeric value. Threshold value.}
}
\value{
A \code{\link[sp]{SpatialPolygons}} object.
}
\description{
Polygon around \var{gridcells} with a value larger than or equal to the
specified threshold
}
\examples{
r <- raster::raster(
  xmn = 0, xmx = 10,
  ymn = 0, ymx = 10,
  crs = "OGC:CRS84",
  resolution = c(1, 1)
)
r <- raster::init(r, fun = "cell")

ip1 <- isoline_from_raster(r, alpha = 45)
ip2 <- isoline_from_raster(r, alpha = 87)

raster::plot(r)
sp::plot(ip1, border = "black", add = TRUE)
sp::plot(ip2, border = "blue", add = TRUE)

}
