% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{ncsw}
\alias{ncsw}
\title{\code{netCDF} functionality suitable for \code{SOILWAT2}}
\arguments{
\item{x}{A character string (file name that will be opened and closed),
an object of class \code{"NetCDF"} from the \code{RNetCDF} package (kept open), or
an object of class \code{"ncdf4"} from the \code{ncdf4} package (closed with a warning;
the file is then opened by its file name and closed on exit).}

\item{long_name}{A character string. The \code{"long_name"} attribute.}

\item{units}{A character string.  The \code{"units"} attribute.}

\item{cell_methods}{A character string. The \code{"cell_methods"} attribute.
Previously, misspelled as \code{"cell_method"}.}

\item{cell_method}{A character string. Passed to \code{"cell_methods"} for
backwards compatibility; will be removed in the version.}

\item{coordinates}{A character string. The \code{"coordinates"} attribute.}

\item{grid_mapping}{A character string. The \code{"grid_mapping"} attribute.}

\item{attributes}{A named vector or named list of character strings.}

\item{dataType}{A character string. A \code{netCDF} data type, see \code{\link[=ncDataType]{ncDataType()}}.}

\item{nameDimX}{A character string. The name of the \code{X}-axis dimension and
coordinate variable.}

\item{nameDimY}{A character string. The name of the \code{Y}-axis dimension and
coordinate variable.}

\item{nameCRS}{A character string. The name of the \code{CRS} variable.}

\item{nameAxis}{A character string. The name of the axis dimension and
associated coordinate variable.}

\item{values}{A numeric object. The values to write a variable if not \code{NULL}.}

\item{siteValues}{A numeric object. The values of the discrete \code{site} axis
variable.}

\item{timeValues}{A numeric vector. The values of the time axis variable.}

\item{verticalValues}{A numeric vector. The values of the vertical axis
variable.}

\item{pftValues}{A vector of character strings. The value of the variable
representing plant functional types.}

\item{var_chunksizes_xyzt}{A numeric vector. The chunk sizes if not \code{NA},
see \code{\link[RNetCDF:var.def.nc]{RNetCDF::var.def.nc()}} for more detail.}

\item{isUnlimitedDim}{A logical value. Define the new dimension as unlimited?}
}
\description{
\code{netCDF} functionality suitable for \code{SOILWAT2}
}
