% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{ncsw}
\alias{ncsw}
\title{\code{netCDF} functionality suitable for \code{SOILWAT2}}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{attributes}{A named vector or named list of character strings.}

\item{dataType}{A character string. A \code{netCDF} data type.}
}
\description{
\code{netCDF} functionality suitable for \code{SOILWAT2}
}
