% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_netCDF.R
\name{read_attributes_from_netCDF}
\alias{read_attributes_from_netCDF}
\title{Read all attributes of a group from a \var{netCDF}}
\usage{
read_attributes_from_netCDF(
  x,
  group = c("var", "xy", "crs", "time", "vertical", "global", "all"),
  var = NULL,
  xy_names = c("lon", "lat"),
  nc_name_crs = "crs"
)
}
\arguments{
\item{x}{An object identifying a \var{netCDF} file, i.e.,
a character string as file name or an object of class \var{ncdf4} derived
from \code{ncdf4::nc_open}.}

\item{group}{A character string. Specifies which attributes to extract.}

\item{var}{A character string. The name or names of the variables from
which to extract attributes, used if the "var" \code{group} is requested.}

\item{xy_names}{A vector with two character strings. The names of the
\var{x} and \var{y} spatial dimensions of the \var{netCDF} file.}

\item{nc_name_crs}{A character string. The name of the \var{crs} variable
in the \var{netCDF}.
Function \code{\link{create_netCDF}} hard codes \var{"crs"}.}
}
\value{
A named list
}
\description{
Read all attributes of a group from a \var{netCDF}
}
\section{Details}{

  Example code is available in the documentation of
  \code{\link{read_netCDF}}.

 \itemize{
  \item If attributes of a variable are requested (group is \var{"var"}),
        then the variable name must be passed to \code{var}.
  \item If \var{xy-space} attributes are requested (group is \var{"xy"}),
        then the names of the x and y dimension must be passed to
        \code{xy_names}.
  \item If \var{crs} attributes are requested (group is \var{"crs"}),
        then the name must be passed to \code{nc_name_crs}.
  \item To obtain all attributes from each group, pass "all" to \code{group};
        \code{var}, \code{xy_names}, \code{nc_name_crs}
        must be correctly specified.
}
}

