% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{renameSpatialNCSW}
\alias{renameSpatialNCSW}
\title{Rename spatial dimensions and variables}
\usage{
renameSpatialNCSW(
  x,
  nameDimX = c(orig = "longitude", to = "lon"),
  nameDimY = c(orig = "latitude", to = "lat"),
  nameCRS = c(orig = "crs", to = "crs_geogsc")
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed),
an object of class \code{"NetCDF"} from the \code{RNetCDF} package (kept open), or
an object of class \code{"ncdf4"} from the \code{ncdf4} package (kept open).}

\item{nameDimX}{A character string. The name of the \code{X}-axis dimension and
coordinate variable.}

\item{nameDimY}{A character string. The name of the \code{Y}-axis dimension and
coordinate variable.}

\item{nameCRS}{A character string. The name of the \code{CRS} variable.}
}
\description{
Rename spatial dimensions and variables
}
