% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{renameSpatialNCSW}
\alias{renameSpatialNCSW}
\title{Rename spatial dimensions and variables}
\usage{
renameSpatialNCSW(
  x,
  nameDimX = c(orig = "longitude", to = "lon"),
  nameDimY = c(orig = "latitude", to = "lat"),
  nameCRS = c(orig = "crs", to = "crs_geogsc")
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}
}
\description{
Rename spatial dimensions and variables
}
