% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setAxisTimeNCSW}
\alias{setAxisTimeNCSW}
\title{Add a time dimension and coordinate variable}
\usage{
setAxisTimeNCSW(
  x,
  startYear,
  timeValues,
  nameAxis = "time",
  calendar = c("standard", "365_day", "366_day"),
  dataType = "NC_DOUBLE"
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{dataType}{A character string. A \code{netCDF} data type.}
}
\description{
Add a time dimension and coordinate variable
}
