% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setAxisVerticalNCSW}
\alias{setAxisVerticalNCSW}
\title{Add a vertical dimension and coordinate variable (\code{Z}-axis)}
\usage{
setAxisVerticalNCSW(
  x,
  verticalValues,
  verticalUpperBound = NULL,
  verticalLowerBound = NULL,
  verticalType = c("values", "layers"),
  nameAxis = "vertical",
  units = "cm",
  dataType = "NC_INT"
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{verticalValues}{A numeric vector. The values of the vertical axis
variable.}

\item{verticalUpperBound}{A numerical vector. The upper (top) soil depths
(bounds) of each soil layer (as defined by \code{verticalValues}).}

\item{verticalLowerBound}{A numerical vector. The lower (bottom) soil depths
(bounds) of each soil layer (as defined by \code{verticalValues}).}

\item{verticalType}{A character string. The type of the vertical axis.}

\item{nameAxis}{A character string. The name of the axis dimension and
associated coordinate variable.}

\item{units}{A character string. The length units of \code{verticalType}.}

\item{dataType}{A character string. A \code{netCDF} data type, see \code{\link[=ncDataType]{ncDataType()}}.}
}
\description{
Add a vertical dimension and coordinate variable (\code{Z}-axis)
}
