% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setSpatialBoundsNCSW}
\alias{setSpatialBoundsNCSW}
\title{Set spatial bounds on x-axis and y-axis}
\usage{
setSpatialBoundsNCSW(
  x,
  nameBnds = "bnds",
  nameBndsX = "lon_bnds",
  nameBndsY = "lat_bnds",
  nameDimX = "longitude",
  nameDimY = "latitude",
  valuesBndsX = NULL,
  valuesBndsY = NULL
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}
}
\description{
Set spatial bounds on x-axis and y-axis
}
