% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{setVariableNCSW}
\alias{setVariableNCSW}
\title{Add a variable}
\usage{
setVariableNCSW(
  x,
  varName,
  values = NULL,
  start = NULL,
  count = NULL,
  var_chunksizes_xyzt = NA,
  dataType = c("NC_DOUBLE", "NC_FLOAT"),
  dimensions = NULL,
  deflate = 5L,
  long_name = NULL,
  units = NULL,
  cell_method = NULL,
  coordinates = paste("latitude", "longitude"),
  grid_mapping = "crs",
  attributes = NULL
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{dataType}{A character string. A \code{netCDF} data type.}

\item{attributes}{A named vector or named list of character strings.}
}
\description{
Add a variable
}
\section{Details}{

\enumerate{
\item Create variable of \code{dataType} and \code{dimensions} if not present.
\item Write values if \code{values}, \code{count}, and optionally \code{start} are provided.
\item Add attributes \code{"long_name"}, \code{"units"}, \code{"cell_method"},
\code{"coordinates"} and \code{"grid_mapping"} if provided.
If \code{"long_name"} is not provided and there is no attribute with that
name, then \code{varName} is used as value for \code{"long_name"}.
\item Add additional attributes if provided.
}
}

