% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{uniqueDifferences}
\alias{uniqueDifferences}
\title{Unique differences}
\usage{
uniqueDifferences(
  x,
  uniqueDifferences = TRUE,
  allowMultiple = FALSE,
  tolerance = sqrt(.Machine[["double.eps"]])
)
}
\description{
Unique differences
}
\examples{
x <- seq(0, 1, by = 1 / 24)
uniqueDifferences(x)

x[[2L]] <- x[[2L]] + sqrt(.Machine[["double.eps"]])
try(uniqueDifferences(x))
uniqueDifferences(x, tolerance = 1e-7)
uniqueDifferences(x, allowMultiple = TRUE)

}
