% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_spatial.R
\name{variogram_range}
\alias{variogram_range}
\title{Calculate range of variogram in kilometers}
\usage{
variogram_range(x, project_to_utm = TRUE, sub_samplepoints_N = NULL)
}
\arguments{
\item{x}{A \code{\link[raster:RasterLayer-class]{raster::RasterLayer}}
in \var{un}-projected coordinates (i.e., longitude/latitude)}
}
\description{
Calculate range of variogram in kilometers
}
