% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{writeTerraToNCSW}
\alias{writeTerraToNCSW}
\title{Write a \code{terra} raster object to a \code{netCDF} file suitable for \code{SOILWAT2}}
\usage{
writeTerraToNCSW(
  x,
  filename,
  var_chunksizes_xyzt = NA,
  dataType = c("double", "float", "integer", "short", "byte", "char"),
  increasingLat = TRUE,
  nameLon = "longitude",
  nameLat = "latitude",
  addSpatialBounds = TRUE,
  nameAxisVertical = "vertical",
  verticalValues = NULL,
  nameAxisTime = "time",
  timeValues = NULL,
  deleteGlobalAttributes = c("created_by", "created_date", "date")
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed) or
an object of class \code{"NetCDF"} from the \code{RNetCDF} package
(an open connection to a \code{NetCDF} dataset that will be kept open).}

\item{dataType}{A character string. A \code{netCDF} data type.}
}
\description{
Write a \code{terra} raster object to a \code{netCDF} file suitable for \code{SOILWAT2}
}
