% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ncSOILWAT2.R
\name{writeTerraToNCSW}
\alias{writeTerraToNCSW}
\title{Write spatial objects to a \code{netCDF} file suitable for \code{SOILWAT2}}
\usage{
writeTerraToNCSW(
  x,
  filename,
  var_chunksizes_xyzt = NA,
  dataType = c("double", "float", "integer", "short", "byte", "char"),
  increasingLat = TRUE,
  nameDimX = "longitude",
  nameDimY = "latitude",
  addSpatialBounds = TRUE,
  nameAxisVertical = "vertical",
  verticalValues = NULL,
  nameAxisTime = "time",
  timeValues = NULL,
  deleteGlobalAttributes = c("created_by", "created_date", "date")
)
}
\arguments{
\item{x}{A character string (file name that will be opened and closed),
an object of class \code{"NetCDF"} from the \code{RNetCDF} package (kept open), or
an object of class \code{"ncdf4"} from the \code{ncdf4} package (kept open).}

\item{filename}{A character string. The file name of the \code{netCDF} file
to be created.}

\item{var_chunksizes_xyzt}{A numeric vector. The chunk sizes if not \code{NA},
see \code{\link[RNetCDF:var.def.nc]{RNetCDF::var.def.nc()}} for more detail.}

\item{dataType}{A character string. A \code{netCDF} data type, see \code{\link[=ncDataType]{ncDataType()}}.}

\item{increasingLat}{A logical value. Sort \code{Y}-axis values increasingly.}

\item{nameDimX}{A character string. The name of the \code{X}-axis dimension and
coordinate variable.}

\item{nameDimY}{A character string. The name of the \code{Y}-axis dimension and
coordinate variable.}

\item{addSpatialBounds}{A logical value. Add spatial bounds.}

\item{nameAxisVertical}{A character string. The name of the \code{Z}-axis
dimension and coordinate variable.}

\item{verticalValues}{A numeric vector. The values of the vertical axis
variable.}

\item{nameAxisTime}{A character string. The name of the \code{T}-axis
dimension and coordinate variable.}

\item{timeValues}{A numeric vector. The values of the time axis variable.}

\item{deleteGlobalAttributes}{A vector of character strings.
Global attributes to delete.}
}
\description{
Write spatial objects to a \code{netCDF} file suitable for \code{SOILWAT2}
}
\section{Details}{

\code{writeTerraToNCSW()} writes a \code{terra} raster object to a \code{netCDF} file
suitable for \code{SOILWAT2} simulations.
}

